no commit message

+++ wxapp.wmj.com.cn/app/api/controller/LockAuth.php
@@ -1,7 +1,7 @@
 <?php 
 /*
  module:		钥匙管理
- create_time:	2020-04-08 22:56:16
+ create_time:	2020-04-27 22:50:31
  author:		
  contact:		
 */
@@ -15,6 +15,98 @@ use think\facade\Log;
 
 class LockAuth extends Common {
 
+
+
+/*start*/
+	/**
+	* @api {post} /LockAuth/getauthdetailbyid 05、查看数据
+	* @apiGroup LockAuth
+	* @apiVersion 1.0.0
+	* @apiDescription  查看数据
+	
+	* @apiParam (输入参数：) {string}     		lockauth_id 主键ID
+
+	* @apiHeader {String} Authorization 用户授权token
+	* @apiHeaderExample {json} Header-示例:
+	* "Authorization: eyJhbGciOiJIUzUxMiJ9.eyJzdWIiOjM2NzgsImF1ZGllbmNlIjoid2ViIiwib3BlbkFJZCI6MTM2NywiY3JlYXRlZCI6MTUzMzg3OTM2ODA0Nywicm9sZXMiOiJVU0VSIiwiZXhwIjoxNTM0NDg0MTY4fQ.Gl5L-NpuwhjuPXFuhPax8ak5c64skjDTCBC64N_QdKQ2VT-zZeceuzXB9TqaYJuhkwNYEhrV3pUx1zhMWG7Org"
+
+	* @apiParam (失败返回参数：) {object}     	array 返回结果集
+	* @apiParam (失败返回参数：) {string}     	array.status 返回错误码 201
+	* @apiParam (失败返回参数：) {string}     	array.msg 返回错误消息
+	* @apiParam (成功返回参数：) {string}     	array 返回结果集
+	* @apiParam (成功返回参数：) {string}     	array.status 返回错误码 200
+	* @apiParam (成功返回参数：) {string}     	array.data 返回数据详情
+	* @apiSuccessExample {json} 01 成功示例
+	* {"status":"200","data":""}
+	* @apiErrorExample {json} 02 失败示例
+	* {"status":"201","msg":"没有数据"}
+	*/
+	function getauthdetailbyid(){
+		$data['lockauth_id'] = $this->request->post('lockauth_id','','intval');
+		if(!$data['lockauth_id']) return json(['status'=>$this->errorCode,'msg'=>'lockauth_id不能为空']);
+		try{
+			$res = checkData(LockAuthDb::query('select a.*,b.headimgurl,b.nickname,b.mobile,c.lock_name from cd_lockauth as a inner join cd_member as b inner join cd_lock as c where a.member_id=b.member_id and a.lock_id=c.lock_id and a.lockauth_id = '.$data['lockauth_id']));
+		}catch (\Exception $e){
+			return json(['status'=>$this->errorCode,'msg'=>$e->getMessage()]);
+		}
+		Log::info('接口输出：'.print_r($res,true));
+		return json(['status'=>$this->successCode,'data'=>$res]);
+	}
+/*end*/
+
+
+/*start*/
+	/**
+	* @api {post} /LockAuth/getauthlistbylockid 08、根据锁id查询钥匙
+	* @apiGroup LockAuth
+	* @apiVersion 1.0.0
+	* @apiDescription  根据锁id查询钥匙
+
+	* @apiHeader {String} Authorization 用户授权token
+	* @apiHeaderExample {json} Header-示例:
+	* "Authorization: eyJhbGciOiJIUzUxMiJ9.eyJzdWIiOjM2NzgsImF1ZGllbmNlIjoid2ViIiwib3BlbkFJZCI6MTM2NywiY3JlYXRlZCI6MTUzMzg3OTM2ODA0Nywicm9sZXMiOiJVU0VSIiwiZXhwIjoxNTM0NDg0MTY4fQ.Gl5L-NpuwhjuPXFuhPax8ak5c64skjDTCBC64N_QdKQ2VT-zZeceuzXB9TqaYJuhkwNYEhrV3pUx1zhMWG7Org"
+
+	* @apiParam (输入参数：) {int}     		[limit] 每页数据条数（默认20）
+	* @apiParam (输入参数：) {int}     		[page] 当前页码
+	* @apiParam (输入参数：) {string}		lock_id 锁id
+    * @apiParam (输入参数：) {int}		auth_status 钥匙状态
+    
+	* @apiParam (失败返回参数：) {object}     	array 返回结果集
+	* @apiParam (失败返回参数：) {string}     	array.status 返回错误码 201
+	* @apiParam (失败返回参数：) {string}     	array.msg 返回错误消息
+	* @apiParam (成功返回参数：) {string}     	array 返回结果集
+	* @apiParam (成功返回参数：) {string}     	array.status 返回错误码 200
+	* @apiParam (成功返回参数：) {string}     	array.data 返回数据
+	* @apiParam (成功返回参数：) {string}     	array.data.list 返回数据列表
+	* @apiParam (成功返回参数：) {string}     	array.data.count 返回数据总数
+	* @apiSuccessExample {json} 01 成功示例
+	* {"status":"200","data":""}
+	* @apiErrorExample {json} 02 失败示例
+	* {"status":" 201","msg":"查询失败"}
+	*/
+	function getauthlistbylockid(){
+		$limit  = $this->request->post('limit', 20, 'intval');
+		$page   = $this->request->post('page', 1, 'intval');
+        $lock_id=$this->request->post('lock_id', '', 'serach_in');
+        $auth_status=$this->request->post('auth_status', '', 'serach_in');
+        if(!$lock_id) return json(['status'=>$this->errorCode,'msg'=>'lock_id不能为空']);
+		$where = [];
+		$where['a.lock_id'] = $lock_id;
+        $where['a.auth_status'] = $auth_status;
+		$limit = ($page-1) * $limit.','.$limit;
+		$field = '*';
+		$orderby = 'lockauth_id desc';
+
+		try{
+			$sql = 'select a.*,b.headimgurl,b.nickname,b.mobile,c.lock_name from cd_lockauth as a inner join cd_member as b inner join cd_lock as c where a.member_id=b.member_id and a.lock_id=c.lock_id';
+			$res = \xhadmin\CommonService::loadList($sql,formatWhere($where),$limit,$orderby);
+		}catch(\Exception $e){
+			return json(['status'=>$this->errorCode,'msg'=>$e->getMessage()]);
+		}
+
+		return json(['status'=>$this->successCode,'data'=>htmlOutList($res)]);
+	}
+/*end*/
 
 
 /*start*/
@@ -55,7 +147,7 @@ class LockAuth extends Common {
 		$isapplywhere['lock_id']=$data['lock_id'];
 		$isapplylock=LockAuthDb::getWhereInfo($isapplywhere,$field);
 		if ($isapplylock) {
-			mlog("isapplylock".json_encode($isapplylock));
+			//mlog("isapplylock".json_encode($isapplylock));
 			return json(['status'=>$this->errorCode,'msg'=>'您已经申请过，请等待审核']);
 		}
 		try {
@@ -77,9 +169,7 @@ class LockAuth extends Common {
 	* @apiHeader {String} Authorization 用户授权token
 	* @apiHeaderExample {json} Header-示例:
 	* "Authorization: eyJhbGciOiJIUzUxMiJ9.eyJzdWIiOjM2NzgsImF1ZGllbmNlIjoid2ViIiwib3BlbkFJZCI6MTM2NywiY3JlYXRlZCI6MTUzMzg3OTM2ODA0Nywicm9sZXMiOiJVU0VSIiwiZXhwIjoxNTM0NDg0MTY4fQ.Gl5L-NpuwhjuPXFuhPax8ak5c64skjDTCBC64N_QdKQ2VT-zZeceuzXB9TqaYJuhkwNYEhrV3pUx1zhMWG7Org"
-	* @apiParam (输入参数：) {int}				lock_id 锁ID 
-	* @apiParam (输入参数：) {string}			member_id 会员ID 
-	* @apiParam (输入参数：) {int}				auth_member_id 分享人会员ID 
+	* @apiParam (输入参数：) {int}				lockauth_id 钥匙ID 
 	* @apiParam (输入参数：) {int}				auth_sharelimit 可分享钥匙数 
 	* @apiParam (输入参数：) {int}				auth_openlimit 开门限制次数 
 	* @apiParam (输入参数：) {string}			auth_starttime 有效期起始时间 
@@ -87,9 +177,7 @@ class LockAuth extends Common {
 	* @apiParam (输入参数：) {int}				auth_isadmin 是否管理员 
 	* @apiParam (输入参数：) {int}				auth_shareability 分享权限 开启|1,关闭|0
 	* @apiParam (输入参数：) {string}			remark 备注 
-	* @apiParam (输入参数：) {string}			create_time 创建时间 
 	* @apiParam (输入参数：) {int}				auth_status 审核状态 已审核|1,未审核|0
-	* @apiParam (输入参数：) {int}				user_id 管理员ID 
 
 	* @apiParam (失败返回参数：) {object}     	array 返回结果集
 	* @apiParam (失败返回参数：) {string}     	array.status 返回错误码  201
@@ -103,7 +191,7 @@ class LockAuth extends Common {
 	* {"status":" 201","msg":"操作失败"}
 	*/
 	function verifyauth(){
-		$postField = 'lockauth_id,lock_id,member_id,auth_member_id,auth_sharelimit,auth_openlimit,auth_starttime,auth_endtime,auth_isadmin,auth_shareability,remark,create_time,auth_status,user_id';
+		$postField = 'lockauth_id,auth_sharelimit,auth_openlimit,auth_starttime,auth_endtime,auth_isadmin,auth_shareability,remark,auth_status';
 		$data = $this->request->only(explode(',',$postField),'post',null);
 		if(empty($data['lockauth_id'])) return json(['status'=>$this->errorCode,'msg'=>'参数错误']);
 		try {
@@ -114,7 +202,10 @@ class LockAuth extends Common {
 		}
 		return json(['status'=>$this->successCode,'msg'=>'操作成功']);
 	}
+/*end*/
 
+
+/*start*/
 	/**
 	* @api {post} /LockAuth/delete 04、删除
 	* @apiGroup LockAuth
@@ -140,48 +231,27 @@ class LockAuth extends Common {
 	function delete(){
 		$idx =  $this->request->post('lockauth_ids', '', 'serach_in');
 		if(empty($idx)) return json(['status'=>$this->errorCode,'msg'=>'参数错误']);
+		
+		//查询是否管理员，是的话不能删除自己
+		$field='lockauth_id,lock_id,member_id,auth_isadmin';
+		$isadminwhere['lockauth_id']=$idx;
+		$isadminlock=LockAuthDb::getWhereInfo($isadminwhere,$field);
+		mlog("delete_auth_isadminlock:".json_encode($isadminlock));
 		try{
-			$data['lockauth_id'] = explode(',',$idx);
-			LockAuthService::delete($data);
+		    if($isadminlock['auth_isadmin'])
+	       	{
+		     mlog("delete_auth_isadminlock_notdeleteself");
+		     return json(['status'=>$this->errorCode,'msg'=>'管理员不能在移动端删除']);
+	    	}
+	    	else
+	    	{
+			 $data['lockauth_id'] = explode(',',$idx);
+			 LockAuthService::delete($data);
+	    	}
 		}catch(\Exception $e){
 			return json(['status'=>$this->errorCode,'msg'=>'操作失败']);
 		}
 		return json(['status'=>$this->successCode,'msg'=>'操作成功']);
-	}
-
-	/**
-	* @api {post} /LockAuth/getauthdetailbyid 05、查看数据
-	* @apiGroup LockAuth
-	* @apiVersion 1.0.0
-	* @apiDescription  查看数据
-	
-	* @apiParam (输入参数：) {string}     		lockauth_id 主键ID
-
-	* @apiHeader {String} Authorization 用户授权token
-	* @apiHeaderExample {json} Header-示例:
-	* "Authorization: eyJhbGciOiJIUzUxMiJ9.eyJzdWIiOjM2NzgsImF1ZGllbmNlIjoid2ViIiwib3BlbkFJZCI6MTM2NywiY3JlYXRlZCI6MTUzMzg3OTM2ODA0Nywicm9sZXMiOiJVU0VSIiwiZXhwIjoxNTM0NDg0MTY4fQ.Gl5L-NpuwhjuPXFuhPax8ak5c64skjDTCBC64N_QdKQ2VT-zZeceuzXB9TqaYJuhkwNYEhrV3pUx1zhMWG7Org"
-
-	* @apiParam (失败返回参数：) {object}     	array 返回结果集
-	* @apiParam (失败返回参数：) {string}     	array.status 返回错误码 201
-	* @apiParam (失败返回参数：) {string}     	array.msg 返回错误消息
-	* @apiParam (成功返回参数：) {string}     	array 返回结果集
-	* @apiParam (成功返回参数：) {string}     	array.status 返回错误码 200
-	* @apiParam (成功返回参数：) {string}     	array.data 返回数据详情
-	* @apiSuccessExample {json} 01 成功示例
-	* {"status":"200","data":""}
-	* @apiErrorExample {json} 02 失败示例
-	* {"status":"201","msg":"没有数据"}
-	*/
-	function getauthdetailbyid(){
-		$data['lockauth_id'] = $this->request->post('lockauth_id','','intval');
-		try{
-			$field='lockauth_id,lock_id,member_id,auth_member_id,auth_endtime,auth_starttime,auth_shareability,remark,create_time';
-			$res  = checkData(LockAuthDb::getWhereInfo($data,$field));
-		}catch (\Exception $e){
-			return json(['status'=>$this->errorCode,'msg'=>$e->getMessage()]);
-		}
-		Log::info('接口输出：'.print_r($res,true));
-		return json(['status'=>$this->successCode,'data'=>$res]);
 	}
 /*end*/
